sched: Remove the unnecessary cast from pid_t to int

+++ binfmt/binfmt_execmodule.c
@@ -294,7 +294,7 @@ int exec_module(FAR struct binary_s *binp,
     }
 #endif
 
-  return (int)pid;
+  return pid;
 
 errout_with_tcbinit:
 #ifndef CONFIG_BUILD_KERNEL
